+++ onnxruntime/python/tools/quantization/E2E_example_model/image_classification/ReadMe.txt
@@ -1,0 +1,2 @@
+call run.py to calibrate, quantize and run the quantized model, e.g.:
+python run.py --input_model mobilenetv2-7.onnx --output_model mobilenetv2-7.quant.onnx --calibrate_dataset ./test_images/
